requirements.txt
pet-pet-gif==1.0.2
pillow==11.2.1
py-cord==2.6.1
PyNaCl==1.5.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2

